refactor: separate codebase - brotli
