+++ index.js
@@ -3,12 +3,13 @@ import readline from 'node:readline'
 import os from 'node:os';
 import * as path from "node:path";
 import fs from "node:fs";
-import * as zlib from "node:zlib";
 import osData from "./os-functions.js";
 import {errorHandler, exit, printCurrentDirectory, selectCommand, SendMessage, startApp} from "./service-functions.js";
 import calculateHash from "./hash.js";
 import copyFile from "./copy.js";
 import moveFile from "./move.js";
+import compressBrotli from "./compress-brotli.js";
+import decompressBrotli from "./decompress-brotli.js";
 
 export const rl = readline.createInterface({input: process.stdin, output: process.stdout});
 const args = process.argv.slice(2);
@@ -58,29 +59,11 @@ rl.on('line', async (mes) => {
             break;
         }
         case 'compress': {
-            const copyFromCompress = mes.replace(command, '').trimStart().split(' ').at(0);
-            const copyToCompress = mes.replace(command, '').trimStart().split(' ').at(1);
-            const brotli = zlib.createBrotliCompress();
-            const compressRead = fs.createReadStream(copyFromCompress);
-            const compressWrite = fs.createWriteStream(
-                path.resolve(copyToCompress, path.basename(copyFromCompress).concat('.br'))
-            );
-            compressRead.on('error', err => errorHandler(err));
-            compressWrite.on('error', err => errorHandler(err));
-            compressRead.pipe(brotli).pipe(compressWrite);
+            compressBrotli(mes, command);
             break;
         }
         case 'decompress': {
-            const copyFromDecompress = mes.replace(command, '').trimStart().split(' ').at(0);
-            const copyToDecompress = mes.replace(command, '').trimStart().split(' ').at(1);
-            const brotli = zlib.createBrotliDecompress();
-            const decompressRead = fs.createReadStream(copyFromDecompress);
-            const decompressWrite = fs.createWriteStream(
-                path.resolve(copyToDecompress, path.basename(copyFromDecompress).replace('.br', ''))
-            );
-            decompressRead.on('error', err => errorHandler(err));
-            decompressWrite.on('error', err => errorHandler(err));
-            decompressRead.pipe(brotli).pipe(decompressWrite);
+            decompressBrotli(mes, command);
             break;
         }
         case 'rm': {
